updated forced tunnelling section
I've updated the forced tunnelling section to include this " Also add the IP address for the Azure Machine Learning service where the resources exist, as this is required for access to Compute Instances.". This is required otherwise you get a timeout when trying to connecto to JupyterLab from the browser.